extract array/map/class allocators from variable-allocator.ts
